+++ starsim/sim.py
@@ -1,6 +1,7 @@
 """
 Define core Sim classes
 """
+import itertools
 import numpy as np
 import sciris as sc
 import starsim as ss
@@ -123,10 +124,12 @@ class Sim(ss.Base):
         return products
 
     @property
-    def module_list(self):
-        """ Return a list of all Module instances (stored in standard places) in the Sim; see `sim.module_dict` for the dict version """
-        out = sc.mergelists(
-            self.modules(),
+    def modules(self):
+        """
+        Return an interator over all Module instances (stored in standard places) in the Sim
+        """
+        for module in itertools.chain(
+            self.custom(),
             self.demographics(),
             self.connectors(),
             self.networks(),
@@ -134,18 +137,13 @@ class Sim(ss.Base):
             self.products(),
             self.diseases(),
             self.analyzers(),
-        )
-        return out
-
-    @property
-    def module_dict(self):
-        """ Return a dictionary of all Module instances; see `sim.module_list` for the list version """
-        return ss.utils.nlist_to_dict(self.module_list)
+            ):
+            yield module
 
     @property
     def label_to_name(self):
         """ Return a mapping from module labels to module names, e.g. {'HIV': 'hiv', 'Pregnancy': 'pregnancy'} """
-        return {mod.label: mod.name for mod in self.module_list}
+        return {mod.label: mod.name for mod in self.modules}
 
     def init(self, force=False, timer=False, **kwargs):
         """
@@ -186,6 +184,85 @@ class Sim(ss.Base):
         if self.diagnostics and self.diagnostics.states is not None: # Need not None since dict is empty at this point
             self.diagnostics.store_states(key='init')
         return self
+
+    def get_module(self, query, die=True, match_case=False):
+        """
+        Retrieve a single module
+
+        This method is guaranteed to return a single module if it succeeds, so can be relied on in places
+        where this is expected to retrieve a specific module from the `Sim` (and retrieving no modules or
+        multiple modules would be unexpected). Use `Sim.get_modules()` to retrieve all matching modules.
+
+        Args:
+            query (type/str): Query input supported by `get_modules` (either a `Type` e.g., `ss.SIR`) or a string that will match module names.
+            die (bool): if True (default), raises an exception unless exactly 1 module is found; else return None if no modules found
+            match_case (bool): if `query` is a string, whether to match case (default False)
+
+        Returns:
+            An ss.Module instance
+        """
+        matches = self.get_modules(query, match_case=match_case)
+        if len(matches) > 1:
+            errormsg = f'Multiple matching modules found for {query}; to retrieve all of them, use `Sim.get_modules()` instead'
+            raise Exception(errormsg)
+        elif not matches:
+            if die:
+                errormsg = f'No matching module found for {query}; set die=False to return None instead'
+                raise KeyError(errormsg)
+            else:
+                return None
+        else:
+            return matches[0]
+
+    def get_modules(self, query=None, match_case=False):
+        """
+        Retrieve modules from the Sim
+
+        Retrieve a module by name or by type. The query term can be the type of module to match
+        e.g., `ss.SIR`, `ss.Disease`, `ss.Intervention`, or a string which is tested against
+        module names. The string can contain a `'*'` at the start or end.
+
+        If the `Sim` is not initialized, this will search over any `ss.Module` instances contained in
+        `Sim.pars`. Note that sim initialization might result in the creation of modules, therefore some modules
+        may be retrieved after initialization but not before (e.g. `diseases='sir'` will not be registered
+        as the module `ss.SIR()` until after initialization).
+
+        Args:
+            query (type/str): The module type (e.g. `ss.SIR`) or a case-insensitive string (e.g. `'sir'`); if None, return all modules
+            match_case (bool): if `query` is a string, whether to match case (default False)
+
+        Returns:
+            A list of `ss.Module` instances. The list will be empty if no modules were found
+        """
+        matches = []
+
+        if self.initialized:
+            modules = self.modules
+        else:
+            modules = sc.search(self.pars, type=ss.Module).values()
+
+        # Return all modules if no query
+        if query is None:
+            return list(modules)
+
+        # Loop over all modules, looking for matches
+        for mod in modules:
+            if isinstance(query, type) and isinstance(mod, query):
+                matches.append(mod)
+            elif isinstance(query, str):
+                name = mod.name
+                if not match_case:
+                    query = query.lower()
+                    name = name.lower()
+                if query.startswith('*') and query.endswith('*') and query[1:-1] in name:
+                    matches.append(mod)
+                elif query.startswith('*') and name.endswith(query[1:]):
+                    matches.append(mod)
+                elif query.endswith('*') and name.startswith(query[:-1]):
+                    matches.append(mod)
+                elif name == query:
+                    matches.append(mod)
+        return matches
 
     def init_time(self):
         """ Time indexing; derived values live in the sim rather than in the pars """
@@ -247,6 +324,8 @@ class Sim(ss.Base):
     def init_module_attrs(self, force=False):
         """ Move initialized modules to the sim """
         module_types = ss.modules.module_types()
+        module_types.append('custom') # Support extra catch-all module category
+
         for attr in module_types:
             orig = getattr(self, attr, None)
             if not force and orig is not None:
@@ -263,7 +342,7 @@ class Sim(ss.Base):
 
     def init_modules_pre(self):
         """ Initialize all the modules with the sim """
-        for mod in self.module_list:
+        for mod in self.modules:
             mod.init_pre(self)
         return
 
@@ -273,7 +352,7 @@ class Sim(ss.Base):
         self.dists.init(obj=self, base_seed=self.pars.rand_seed, force=True)
 
         # Copy relevant dists to each module
-        for mod in self.module_list:
+        for mod in self.modules:
             self.dists.copy_to_module(mod)
         return
 
@@ -284,13 +363,13 @@ class Sim(ss.Base):
 
     def init_modules_post(self):
         """ Initialize values in other modules, including networks and time parameters, and do any other post-processing """
-        for mod in self.module_list:
+        for mod in self.modules:
             mod.init_post()
         return
 
     def reset_time_pars(self, force=True):
         """ Reset the time parameters in the modules; used for imposing the sim's timestep on the modules """
-        for mod in self.module_list:
+        for mod in self.modules:
             mod.init_time_pars(force=force)
         return
 
@@ -441,8 +520,8 @@ class Sim(ss.Base):
         self.finalize_results()
 
         # Finalize each module, including the results
-        for module in self.module_list:
-            module.finalize()
+        for mod in self.modules:
+            mod.finalize()
 
         # Resets verbose if needed
         if hasattr(self, '_orig_verbose'):
@@ -558,14 +637,14 @@ class Sim(ss.Base):
                     dist.shrink()
 
             # Finally, shrink the modules
-            for mod in sim.module_list:
+            for mod in sim.modules:
                 with sc.tryexcept(die=die):
                     mod.shrink()
 
             # Check that the module successfully shrunk
             if size_limit:
                 max_size = size_limit*(len(sim)+intercept) # Maximum size in KB
-                for mod in sim.module_list:
+                for mod in sim.modules:
                     size = sc.checkmem(mod, descend=0).bytesize[0]/1e3 # Size in KB
                     if size > max_size:
                         errormsg = f'Module {mod.name} did not successfully shrink: {size:n} KB > {max_size:n} KB; use die=False to turn this message into a warning, or change size_limit to a larger value'
@@ -616,7 +695,7 @@ class Sim(ss.Base):
                 sc.pp(self._call_required)
 
             missing = []
-            for mod in self.module_list:
+            for mod in self.modules:
                 modmissing = mod.check_method_calls()
                 if modmissing:
                     missing.append([type(mod), modmissing])
@@ -867,7 +946,7 @@ class Sim(ss.Base):
                 # Plot data
                 if df is not None:
                     mod = res.module
-                    if mod not in self.module_dict and mod in self.label_to_name: # res.module stores the label; resolve to the name
+                    if mod in self.label_to_name: # res.module stores the label; resolve to the name
                         mod = self.label_to_name[mod]
                     name = res.name
                     found = False